lightrag-mcp/requirements.txt
# Core dependencies from pyproject.toml
mcp>=1.2.0
httpx>=0.28.1
pydantic>=2.11,<3.0.0
pydantic-core>=2.14.0,<3.0.0
python-dotenv>=1.0.1
attrs>=25.3.0

# Additional server dependencies
# fastapi>=0.68.0
# uvicorn>=0.15.0
# requests>=2.26.0
# python-multipart>=0.0.5

# # Development dependencies
# pytest>=6.2.5
# pytest-cov>=2.12.1
# black>=21.7b0
# isort>=5.9.3
# mypy>=0.910
# flake8>=3.9.2
